+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Changes
     * Added `token_data` config option to OpenmaizeJWT. This allows the developer to add extra data to the token.
+    * Changed the way the confirm option is handled in the mix generator.
 
 ## v0.18.0